requirements.txt
pydantic 
pydantic-settings 
websockets
fastapi
uvicorn
pyaudio
dashscope
mss
pyside6
lancedb
sentence_transformers
openai
pyautogui
pyvts
pyperclip
blivedm
pillow
qasync
playwright
json5